refactor(receive): rename receive slice module
